RESEARCH-BOUNDARY PREP — the two code modules leave the capture directories; five dead evidence citations resolve.
Split-independent hygiene from the product/research split design (P3 §2.4 C1 + §6.6).
The split itself is NOT decided; nothing here depends on it.

C1 — code out of research dirs (behaviour byte-identical, git mv):
  examples/polaris/scripts/lib-css.ts      → extract/computed/lib-css.ts
  extract/fidelity-matrix/scripts/env.ts   → extract/figma/visual-parity/env.ts
  Re-export shims stay at both old paths. Importers repointed: evals/run.ts,
  parity/snapshot-rest.ts, extract/figma/visual-parity/figma-api.ts (now ./env.js),
  examples/polaris/scripts/promote.ts. No eval id changes; golden.json untouched.

Five dead evidence citations (no number changed):
  docs/20  extract/computed/out/badge/decisions.json — removed in d10511c4; rows
           survive in extract/computed/out/astryx/badge/decisions.json (now cited)
  docs/23  examples/astryx/out/code-extraction.json — gitignored extraction output;
           regenerate command named (table row + both bash blocks; docs/22 block too)
  docs/22+23  examples/mui/.mui-sandbox/… — gitignored install sandbox, PROVENANCE cited
  docs/21  examples/carbon/.carbon-sandbox — gitignored, recipe recreates it
  README   extract/computed/out/acme — `acme` declared a stand-in; no such path in-tree

Why docs:check missed them: its link pass resolves only markdown [text](target)
links; backtick citations and paths inside fenced shell blocks are never resolved.
A backtick pass today would red 28 more lines (gitignored build outputs,
placeholders, historical script names), so it is not a one-line fix — see PR.
