+++ docs/24-what-works.md
@@ -534,7 +534,7 @@ npm run capability:fresh
 | artifact | sha256 (12) | bytes | what it supplied |
 |---|---|---|---|
 | `conformance/MANIFEST.json` | `8231a2e8b195` | 76,456 | CSS/DOM frontier vocabulary |
-| `docs/22-generality.md` | `65eb9fda0892` | 72,566 | coverage denominators (docs/22 §8.3 table) |
+| `docs/22-generality.md` | `18e87d11b569` | 72,885 | coverage denominators (docs/22 §8.3 table) |
 | `evals/golden.json` | `9a2856a58666` | 32,256 | generated-source golden manifest |
 | `evals/results.json` | `5081770dbb7e` | 6,982 | executable claim suite (registry ids + size; the pass column is the suite's own output) |
 | `examples/untitled-ui/renders/fidelity.json` | `0a468d6682bf` | 84,415 | Untitled UI scored fidelity table |
