+++ extract/CERTIFICATION.md
@@ -70,7 +70,7 @@
 ## What the certification caught (fixed at `331275e`)
 
 1. **Stale committed catalog** (#39): `ds.token` went 1.1.0 (the live size scale) without the catalog being re-emitted — `catalog.json`, `index.json`, and `components/token.json` were behind their own source. Regenerated and committed. **Named residue:** catalog staleness has no standing guard (nothing fails when a contract changes and the catalog doesn't) — and the catalog embeds provenance (`catalogVersion`, `gitCommit`), so it can only ever byte-reproduce at the commit it was generated from. A catalog-freshness eval that compares everything *except* the provenance block is the identified fix.
-2. **Hardcoded owner path in `env.ts`**: `extract/fidelity-matrix/scripts/env.ts` fell back to `/Users/tjpitre/Sites/ds-contracts-poc/.env.local` by absolute path — the certification's first "no token" probe *silently succeeded by reading the owner's token*. Now resolved portably (cwd `.env.local`, else the git main checkout's via `--git-common-dir`); a tokenless fresh clone fails by name (`FIGMA_TOKEN not found (env or .env.local)`), verified in both directions.
+2. **Hardcoded owner path in `env.ts`**: `extract/fidelity-matrix/scripts/env.ts` (since relocated to `extract/figma/visual-parity/env.ts`; the old path is a re-export shim) fell back to `/Users/tjpitre/Sites/ds-contracts-poc/.env.local` by absolute path — the certification's first "no token" probe *silently succeeded by reading the owner's token*. Now resolved portably (cwd `.env.local`, else the git main checkout's via `--git-common-dir`); a tokenless fresh clone fails by name (`FIGMA_TOKEN not found (env or .env.local)`), verified in both directions.
 3. **macOS-only Chromium resolver**: `chromiumExecutable()` knew only the macOS playwright cache and `/Applications` Chrome — on Linux, even `npx playwright install chromium` would not have been found. Linux cache paths and `/usr/bin` fallbacks added; README's quick start now names the two Chromium-driven checks and both escape hatches.
 
 ## Stranger-readiness verdict
